requirements.txt
opencv-python
tqdm
mlflow
matplotlib
dotenv
pycocotools
torch
torchvision
fiftyone
hydra-core
pydantic
seaborn
sam2
